setup .env and server
