+++ .env.example
@@ -1,0 +1,6 @@
+
+# The host the server is listening on
+SERVER_HOST=
+
+# The port number the server is listening on
+SERVER_PORT=
